+++ packages/at_client/test/legacy_self_key_decryption_test.dart
@@ -55,12 +55,6 @@ void main() {
     SelfKeyDecryption selfKeyDecryption = SelfKeyDecryption(mockAtClient);
     SymmetricKey selfEncryptionKey = AESKey.generate(32);
 
-    AtChopsKeys atChopsKeys = AtChopsKeys.create(null, null);
-    atChopsKeys.selfEncryptionKey = selfEncryptionKey;
-
-    AtChops atChopsImpl = AtChopsImpl(atChopsKeys);
-    when(() => mockAtClient.atChops).thenAnswer((_) => atChopsImpl);
-
     var location = 'san francisco';
     var encryptedLocation =
         EncryptionUtil.encryptValue(location, selfEncryptionKey.key);
@@ -79,13 +73,6 @@ void main() {
   test('test to check self encryption key decrypt method with IV', () async {
     SelfKeyDecryption selfKeyDecryption = SelfKeyDecryption(mockAtClient);
     SymmetricKey selfEncryptionKey = AESKey.generate(32);
-
-    AtChopsKeys atChopsKeys = AtChopsKeys.create(null, null);
-    atChopsKeys.selfEncryptionKey = selfEncryptionKey;
-
-    AtChops atChopsImpl = AtChopsImpl(atChopsKeys);
-
-    when(() => mockAtClient.atChops).thenAnswer((_) => atChopsImpl);
 
     var location = 'new york';
     var ivBase64String = 'YmFzZTY0IGVuY29kaW5n';