+++ packages/at_client/test/pairwise_secret_sharing_test.dart
@@ -16,6 +16,7 @@ import 'package:test/test.dart';
 import 'fake_enrollment_directory.dart';
 import 'test_utils/mocks.dart';
 import 'test_utils/test_keypairs.dart';
+import 'test_utils/ml_dsa_keyfile.dart';
 
 class MockAtClient extends Mock implements AtClient {}
 
@@ -83,9 +84,8 @@ void main() {
   MockAtClient buildMockClient(String enrollmentId,
       {MockSyncService? syncService}) {
     final atClient = MockAtClient();
-    final atChops = AtChopsImpl(
-        AtChopsKeys.create(null, pkamKeyPairFor(atSign, enrollmentId)));
-    when(() => atClient.atChops).thenReturn(atChops);
+    when(() => atClient.atKeysIo).thenReturn(keysHoldingApkam(
+        atSign, enrollmentId, pkamKeyPairFor(atSign, enrollmentId)));
     when(() => atClient.getCurrentAtSign()).thenReturn(atSign);
     final sync = syncService ?? MockSyncService();
     when(() => atClient.syncService).thenReturn(sync);